Moving all Zeros to the End of the List
Approach --
Two pointer strategy with 'last_non_zero' pointing to the position where the next non-zero element should go.
How it Iterates --
a. When a non-zero element is at index i, it is swaped with the elemtnat 'last_non_zero'
b. Then 'last_non_zero' is moved forward
This ensures that all non-zero elements are moved tot he front in their original order, while zeros shift toward the end

Time Complexity --
a. The algorithm traverses the list exactly once.
b. for each of the n elements, it perofrms a constant-time operation (comaprison and occassional swap).
c. This results in total of n operations, so the overall time complexity is O(n).